* 'autorestv3' of https://github.com/Azure/autorest.python:
  ChangeLog
  updated to testserver .19 (#529)
  Fix default value enum py3 (#532)
  switiching version.py -> _version.py if we want to keep version file (#522)
  Update changelog (#526)
  removed azure- prefix from ua if there (#523)
  Declare kwargs reserved (#525)
  updating to m4 4.12.276 (#524)
  Add self as reserved keyword (#315)
  Freeze master to 2.7.8 (#258)
  No async flag (#95)

+++ test/vanilla/Expected/AcceptanceTests/Url/url/aio/operations_async/_paths_operations_async.py
@@ -148,7 +148,7 @@ class PathsOperations:
         # Construct URL
         url = self.get_int_one_million.metadata['url']  # type: ignore
         path_format_arguments = {
-            'intPath': self._serialize.url("int_path", int_path, 'float'),
+            'intPath': self._serialize.url("int_path", int_path, 'int'),
         }
         url = self._client.format_url(url, **path_format_arguments)
 
@@ -192,7 +192,7 @@ class PathsOperations:
         # Construct URL
         url = self.get_int_negative_one_million.metadata['url']  # type: ignore
         path_format_arguments = {
-            'intPath': self._serialize.url("int_path", int_path, 'float'),
+            'intPath': self._serialize.url("int_path", int_path, 'int'),
         }
         url = self._client.format_url(url, **path_format_arguments)
 
@@ -236,7 +236,7 @@ class PathsOperations:
         # Construct URL
         url = self.get_ten_billion.metadata['url']  # type: ignore
         path_format_arguments = {
-            'longPath': self._serialize.url("long_path", long_path, 'float'),
+            'longPath': self._serialize.url("long_path", long_path, 'long'),
         }
         url = self._client.format_url(url, **path_format_arguments)
 
@@ -280,7 +280,7 @@ class PathsOperations:
         # Construct URL
         url = self.get_negative_ten_billion.metadata['url']  # type: ignore
         path_format_arguments = {
-            'longPath': self._serialize.url("long_path", long_path, 'float'),
+            'longPath': self._serialize.url("long_path", long_path, 'long'),
         }
         url = self._client.format_url(url, **path_format_arguments)
