+++ test/vanilla/Expected/AcceptanceTests/Url/url/aio/operations_async/_queries_operations_async.py
@@ -188,7 +188,7 @@ class QueriesOperations:
 
         # Construct parameters
         query_parameters = {}  # type: Dict[str, Any]
-        query_parameters['intQuery'] = self._serialize.query("int_query", int_query, 'float')
+        query_parameters['intQuery'] = self._serialize.query("int_query", int_query, 'int')
 
         # Construct headers
         header_parameters = {}  # type: Dict[str, Any]
@@ -229,7 +229,7 @@ class QueriesOperations:
 
         # Construct parameters
         query_parameters = {}  # type: Dict[str, Any]
-        query_parameters['intQuery'] = self._serialize.query("int_query", int_query, 'float')
+        query_parameters['intQuery'] = self._serialize.query("int_query", int_query, 'int')
 
         # Construct headers
         header_parameters = {}  # type: Dict[str, Any]
@@ -314,7 +314,7 @@ class QueriesOperations:
 
         # Construct parameters
         query_parameters = {}  # type: Dict[str, Any]
-        query_parameters['longQuery'] = self._serialize.query("long_query", long_query, 'float')
+        query_parameters['longQuery'] = self._serialize.query("long_query", long_query, 'long')
 
         # Construct headers
         header_parameters = {}  # type: Dict[str, Any]
@@ -355,7 +355,7 @@ class QueriesOperations:
 
         # Construct parameters
         query_parameters = {}  # type: Dict[str, Any]
-        query_parameters['longQuery'] = self._serialize.query("long_query", long_query, 'float')
+        query_parameters['longQuery'] = self._serialize.query("long_query", long_query, 'long')
 
         # Construct headers
         header_parameters = {}  # type: Dict[str, Any]
