+++ test/vanilla/Expected/AcceptanceTests/Url/url/operations/_paths_operations.py
@@ -151,7 +151,7 @@ class PathsOperations(object):
         # Construct URL
         url = self.get_int_one_million.metadata['url']  # type: ignore
         path_format_arguments = {
-            'intPath': self._serialize.url("int_path", int_path, 'float'),
+            'intPath': self._serialize.url("int_path", int_path, 'int'),
         }
         url = self._client.format_url(url, **path_format_arguments)
 
@@ -196,7 +196,7 @@ class PathsOperations(object):
         # Construct URL
         url = self.get_int_negative_one_million.metadata['url']  # type: ignore
         path_format_arguments = {
-            'intPath': self._serialize.url("int_path", int_path, 'float'),
+            'intPath': self._serialize.url("int_path", int_path, 'int'),
         }
         url = self._client.format_url(url, **path_format_arguments)
 
@@ -241,7 +241,7 @@ class PathsOperations(object):
         # Construct URL
         url = self.get_ten_billion.metadata['url']  # type: ignore
         path_format_arguments = {
-            'longPath': self._serialize.url("long_path", long_path, 'float'),
+            'longPath': self._serialize.url("long_path", long_path, 'long'),
         }
         url = self._client.format_url(url, **path_format_arguments)
 
@@ -286,7 +286,7 @@ class PathsOperations(object):
         # Construct URL
         url = self.get_negative_ten_billion.metadata['url']  # type: ignore
         path_format_arguments = {
-            'longPath': self._serialize.url("long_path", long_path, 'float'),
+            'longPath': self._serialize.url("long_path", long_path, 'long'),
         }
         url = self._client.format_url(url, **path_format_arguments)
